remove internal storage
