+++ project/vcmi-app/src/main/AndroidManifest.xml
@@ -26,9 +26,6 @@
             </intent-filter>
         </activity>
         <activity
-            android:name=".ActivityStorage"
-            android:screenOrientation="landscape" />
-        <activity
             android:name=".ActivityError"
             android:screenOrientation="landscape" />
         <activity
