+++ project/vcmi-app/src/main/java/eu/vcmi/vcmi/ServerService.java
@@ -31,7 +31,6 @@ public class ServerService extends Service
     @Override
     public int onStartCommand(Intent intent, int flags, int startId)
     {
-        Storage.initStorage(this);
         NativeMethods.setupCtx(ServerService.this);
         LibsLoader.loadServerLibs();
         if (INTENT_ACTION_KILL_SERVER.equals(intent.getAction()))
