+++ project/vcmi-app/src/main/res/values/strings.xml
@@ -23,26 +23,10 @@
     <string name="launcher_btn_pointermulti_subtitle">Currently: %1$s</string>
     <string name="launcher_btn_sound_title">Sound volume</string>
     <string name="launcher_btn_music_title">Music volume</string>
-    <string name="launcher_btn_storage">Setup storage</string>
-    <string name="launcher_btn_storage_description">Setup where to find VCMI data files</string>
-    <string name="storage_description">VCMI needs original Heroes of Might and Magic 3 data to run properly. You need to install a licensed Heroes of Might and Magic 3 GOG edition on your PC, open it in file explorer and copy Data, Mp3, Maps folders to your Android device. Also you may want to add some VCMI addons with new towns and other nice features.</string>
-    <string name="storage_description2">You have two options to provide this data to VCMI. Either use external storage where specific folder will be created by the game and you need to put the data there. Or use internal storage visible only to VCMI and import the data in form of compressed zip archive.</string>
-    <string name="storage_btn_import_zip">Import VCMI files</string>
-    <string name="storage_btn_import_zip_description">Import HOMM3 files, vcmi mods, saves and settings from a zip file to the internal storage</string>
-    <string name="storage_btn_clear_internal">Clear internal storage</string>
-    <string name="storage_btn_clear_internal_description">Delete HOMM3 data, mods, saves and settings from game internal storage</string>
-    <string name="storage_btn_export_saves">Export saves to external storage</string>
-    <string name="storage_btn_export_saves_description">Export saves to an external storage so that they are not lost. To include them back you can add them to vcmi</string>
-    <string name="storage_btn_use_internal_storage">Use internal storage for VCMI files</string>
-    <string name="storage_btn_use_internal_storage_description">Use app internal storage for VCMI files. Requires no extra permissions. Imported files are visible only to VCMI.</string>
-    <string name="storage_btn_use_external_storage">Use external storage for VCMI files</string>
-    <string name="storage_btn_use_external_storage_description">Use vcmi-data folder in external storage. Will ask storage permissions. May be slow on the latest android devices.</string>
-    <string name="storage_btn_test_external_storage">Test external storage VCMI files</string>
-    <string name="storage_btn_test_external_storage_description">vcmi-data folder should contain Data, Mp3, Maps folders copied from the original game.</string>
     <string name="launcher_version">Current launcher version: %1$s</string>
     <string name="launcher_error_vcmi_data_root_created">VCMI created folder \'%1$s\' (in %2$s). Place your H3 data files there (so that you have a folder %1$s/Data and %1$s/Mp3 etc) and restart the application.</string>
     <string name="launcher_error_vcmi_data_root_failed">Could not create VCMI data folder in %1$s.</string>
-    <string name="launcher_error_h3_data_missing">Could not find data folder in \'%1$s\'. Place your H3 data files there (so that you have a folder %2$s/Data and %2$s/Mp3 etc) and restart the application.</string>
+    <string name="launcher_error_h3_data_missing">Could not find data folder in \'%1$s\'. Place your H3 data files there (so that you have a folder %1$s/Data and %1$s/Mp3 etc) and restart the application.</string>
     <string name="launcher_error_vcmi_data_internal_missing">Could not find or extract vcmi data from app resources. Try to reinstall the app.</string>
     <string name="launcher_error_vcmi_data_internal_update">Could not update vcmi data from app resources. Try to reinstall the app.</string>
     <string name="launcher_error_permissions">This application needs write permissions to use the content stored in external storage</string>
